Rename directory 'context' to 'contexts'
