+++ src/index.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import * as Sentry from '@sentry/browser';
 
-import { CurrentUserProvider } from 'context/CurrentUserContext';
+import { CurrentUserProvider } from 'contexts/CurrentUserContext';
 
 import GlobalStyle from 'styles/GlobalStyle';
